add comments and helper functions
